backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pdfplumber==0.11.0
pymupdf==1.24.2
python-docx==1.1.2
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
openai==1.30.1
qdrant-client==1.9.1
supabase==2.4.5
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
reportlab==4.2.0
weasyprint==62.0
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.7
